Decompose auth middleware and helpers
